Move memory into agent/mason/memory.py with call-time memory_tools()
Removes the import-time env gate; create_agent composes [*all_tools(), *memory_tools()].
memory_tools() returns [] unless AGENT_MEMORY_STORE is set, so the model is only offered
memory tools when configured. Swap the import for a databricks-openai helper when it ships.
